+++ ibformers/data/collators/collators/image.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass
 from typing import ClassVar, List, Optional
 
+import numpy as np
 import torch
 
 from ibformers.data.collators.collators.base import BaseCollator, CollatorABC
@@ -48,7 +49,7 @@ class DetrSubImageImageExtractor(CollatorABC):
         table_images = []
         for feature in features:
             subimage_bbox = feature["structure_image_bbox"]
-            page_image = feature["images"][feature["record_table_page_no"]]
+            page_image = feature["images"][0].astype(np.float32) / 255.0
             table_image = page_image[:, subimage_bbox[1] : subimage_bbox[3], subimage_bbox[0] : subimage_bbox[2]]
             page_images.append(page_image)
             table_images.append(table_image)